Services improved with url checks in UI, missing files FHIR transaction and import supported for R4 and R5

+++ EHR-Sandbox-Ui/src/app/core/_services/_fhir/fhir-bulk.service.ts
@@ -7,6 +7,8 @@ import { TenantService } from '../tenant.service';
 import { ImmunizationRegistryService } from 'src/app/core/_services/immunization-registry.service';
 import { Identifier } from 'fhir/r5';
 import { VaccinationEvent } from 'src/app/core/_model/rest';
+import { SnackBarService } from '../snack-bar.service';
+import { IdUrlVerifyingService } from '../_abstract/id-url-verifying.service';
 
 const httpOptions = {
   headers: new HttpHeaders({ 'Content-Type': 'application/json' })
@@ -17,50 +19,76 @@ const httpOptions = {
 /**
  * Fhir service interacting with the API to parse and serialize resources, and interact with IIS's
  */
-export class FhirBulkService {
+export class FhirBulkService extends IdUrlVerifyingService {
 
-  constructor(private http: HttpClient,
+  constructor(
+    snackBarService: SnackBarService,
+    private http: HttpClient,
     private settings: SettingsService,
     private facilityService: FacilityService,
     private tenantService: TenantService,
-    private registryService: ImmunizationRegistryService) { }
+    private registryService: ImmunizationRegistryService) {
+    super(snackBarService)
+  }
 
 
   groupExportSynch(groupId: string, paramsString: string): Observable<string> {
     const registryId = this.registryService.getCurrentId()
+    const tenantId = this.tenantService.getCurrentId()
+    if (this.idsNotValid(tenantId)) {
+      return of("")
+    }
     return this.http.get(
-      `${this.settings.getApiUrl()}/registry/${registryId}/Group/${groupId}/$export-synch?${paramsString}`,
+      `${this.settings.getApiUrl()}/tenants/${tenantId}/registry/${registryId}/Group/${groupId}/$export-synch?${paramsString}`,
       {
+        ...httpOptions,
         responseType: 'text',
       });
+
   }
 
   groupExportAsynch(groupId: string, paramsString: string): Observable<string> {
     const registryId = this.registryService.getCurrentId()
+    const tenantId = this.tenantService.getCurrentId()
+    if (this.idsNotValid(tenantId)) {
+      return of("")
+    }
     return this.http.get(
-      `${this.settings.getApiUrl()}/registry/${registryId}/Group/${groupId}/$export-asynch?${paramsString}`,
+      `${this.settings.getApiUrl()}/tenants/${tenantId}/registry/${registryId}/Group/${groupId}/$export-asynch?${paramsString}`,
       {
+        ...httpOptions,
         responseType: 'text',
       });
+
+
   }
 
   groupExportStatus(contentUrl: string): Observable<string> {
     const registryId = this.registryService.getCurrentId()
+    const tenantId = this.tenantService.getCurrentId()
+    if (this.idsNotValid(tenantId)) {
+      return of("")
+    }
     return this.http.get(
-      `${this.settings.getApiUrl()}/registry/${registryId}/$export-status`,
+      `${this.settings.getApiUrl()}/tenants/${tenantId}/registry/${registryId}/$export-status`,
       {
+        ...httpOptions,
         responseType: 'text',
-        params: {
-          contentUrl: contentUrl
-        }
       });
+
+
   }
 
   groupExportDelete(contentUrl: string): Observable<string> {
     const registryId = this.registryService.getCurrentId()
+    const tenantId = this.tenantService.getCurrentId()
+    if (this.idsNotValid(tenantId)) {
+      return of("")
+    }
     return this.http.delete(
-      `${this.settings.getApiUrl()}/registry/${registryId}/$export-status`,
+      `${this.settings.getApiUrl()}/tenants/${tenantId}/registry/${registryId}/$export-status`,
       {
+        ...httpOptions,
         responseType: 'text',
         params: {
           contentUrl: contentUrl
@@ -70,32 +98,36 @@ export class FhirBulkService {
 
   groupNdJson(contentUrl: string, loadInFacility: boolean): Observable<string> {
     const registryId = this.registryService.getCurrentId()
+    const tenantId = this.tenantService.getCurrentId()
+    if (this.idsNotValid(tenantId)) {
+      return of("")
+    }
     if (loadInFacility) {
       const facilityId = this.facilityService.getCurrentId()
-      if (facilityId > -1) {
-        return this.http.get(
-          `${this.settings.getApiUrl()}/registry/${registryId}/$export-result`,
-          {
-            responseType: 'text',
-            params: {
-              contentUrl: contentUrl,
-              loadInFacility: facilityId
-            }
-          });
-      } else {
+      if (this.idsNotValid(facilityId)) {
         return of("")
       }
+      return this.http.get(
+        `${this.settings.getApiUrl()}/tenants/${tenantId}/registry/${registryId}/$export-result`,
+        {
+          ...httpOptions,
+          responseType: 'text',
+          params: {
+            contentUrl: contentUrl,
+            loadInFacility: facilityId
+          }
+        });
     } else {
       return this.http.get(
-        `${this.settings.getApiUrl()}/registry/${registryId}/$export-result`,
+        `${this.settings.getApiUrl()}/tenants/${tenantId}/registry/${registryId}/$export-result`,
         {
+          ...httpOptions,
           responseType: 'text',
           params: {
             contentUrl: contentUrl
           }
         });
     }
-
   }
 
   loadNdJson(body: string): Observable<string> {
